Fixed bug
It is possible to move a ticket from one project to another by altering
the URL on the edit ticket page. This should not be allowed. Here
updating project id of ticket causes that problem, because it is only
necessary for create. I'd prefer separate forms for create and update
pages. So I created new form for create which inherits from update form.