Added re-ordering projects list
On the project list page, projects that the user has assigned tickets on
should be shown above the other projects. It is achieved simply by
creating two lists, one for user's projects, another one for other's
projects and merging them. users_tickets are converted to a set and
compared with all project's tickets. AND operator is used to compare two
sets to find common tickets.